requirements.txt
dnspython
dnslib
cryptography
crypto
pycryptodome